Get async tests working (#19)
I've pulled some of the work out of #18 because it was focused on
general test fixes rather than the new uploading feature. Previously,
the test suite was passing because many of the tests were being skipped.
By setting `asyncio_mode` to auto these tests will run. Another very
important change is adding the `auto_enable_custom_integrations` fixture
to `conftest.py`. Without this the custom component won't get setup
properly in the test environment. With the tests actually running now
instead of being skipped, I updated the config flow tests to match the
schema that the integration wants.

I've removed the api test because [Home Assistant advises not to test
details][1], which I interpret as not unit testing that class and
instead having integration tests that make sure that things work as
expected when the integration is connected to Home Assistant.

[1]:
https://developers.home-assistant.io/docs/development_testing/#writing-tests-for-integrations