Implement backup listing (#22)
In order to implement backup listing I've also changed how uploading
works so that we store a flattened version of the metadata so that we
can use it to create out `AgentBackup`s for the list. The `--metadata`
flag for `uplink` doesn't accept nested JSON, so this also adds the
`json_flatten` dependency.